+++ src/components/Menu/index.js
@@ -1,9 +1,13 @@
 import Menu from './Menu';
+import MenuItem from './MenuItem';
+import MenuDivider from './MenuDivider';
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Menu);
+    installPlugin(Vue, MenuItem);
+    installPlugin(Vue, MenuDivider);
   },
 };
 
@@ -13,4 +17,6 @@ export default VuePlugin;
 
 export {
     Menu,
+    MenuItem,
+    MenuDivider,
 };