+++ src/components/Navbar/index.js
@@ -1,9 +1,11 @@
 import Navbar from './Navbar';
+import NavbarSection from './NavbarSection';
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Navbar);
+    installPlugin(Vue, NavbarSection);
   },
 };
 
@@ -13,4 +15,5 @@ export default VuePlugin;
 
 export {
     Navbar,
+    NavbarSection,
 };