+++ src/components/Step/index.js
@@ -1,9 +1,11 @@
 import Step from './Step';
+import StepItem from './StepItem';
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Step);
+    installPlugin(Vue, StepItem);
   },
 };
 
@@ -13,4 +15,5 @@ export default VuePlugin;
 
 export {
     Step,
+    StepItem,
 };