+++ src/components/Tile/index.js
@@ -1,9 +1,20 @@
 import Tile from './Tile';
+import TileAction from './TileAction';
+import TileContent from './TileContent';
+import TileIcon from './TileIcon';
+import TileSubtitle from './TileSubtitle';
+import TileTitle from './TileTitle';
+
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Tile);
+    installPlugin(Vue, TileAction);
+    installPlugin(Vue, TileContent);
+    installPlugin(Vue, TileIcon);
+    installPlugin(Vue, TileSubtitle);
+    installPlugin(Vue, TileTitle);
   },
 };
 
@@ -12,5 +23,10 @@ usePlugin(VuePlugin);
 export default VuePlugin;
 
 export {
-    Tile,
+  Tile,
+  TileAction,
+  TileContent,
+  TileIcon,
+  TileSubtitle,
+  TileTitle,
 };